moved db related methods to TribeDb module
to be able to access methods in all suites
